list a11y: Fix parent-relative position
So far, VCLXAccessibleList::implGetBounds
(which is supposed to report the bounds relative
to the direct parent) and VCLXAccessibleList::getLocationOnScreen
(which is supposed to report the absolte screen position)
were mostly duplicating the logic.

VCLXAccessibleList::getLocationOnScreen seems correct,
but VCLXAccessibleList::implGetBounds reported an
incorrect position for an expanded combobox list.

This is because VCLListBoxHelper::GetDropDownPosSizePixel
was returning a screen position, not a position relative
to the parent.

Fix this by:

* Making VCLListBoxHelper::GetDropDownPosSizePixel return
  a parent-relative position instead.
* Adjusting VCLXAccessibleList::getLocationOnScreen to
  not duplicate most of the logic, but calling
  VCLXAccessibleList::implGetBounds to get the
  parent-relative position and then adding the
  parent's screen position to get the own screen
  position.

The fact that the parent-relative position is incorrect
was noticed while working on fixing the parent-relative
position reported by VCLXAccessibleListItem and even
with an issue in the implementation there, the incorrect
position was still reported.
(The VCLXAccessibleListItem will be done in a separate
upcoming commit.)

Change-Id: I322327753ceec7bc7912ec13c45e208a79dd6976
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181431
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins