list item a11y: Fix parent-relative position
XAccessibleComponent::getLocation etc. are supposed
to return the objects position relative to its direct
accessible parent.

However, IComboListBoxHelper::GetBoundingRectangle used by the
VCLXAccessibleListItem::implGetBounds helper method returns
the position relative to the ComboBox or ListBox for which the
a11y implementation is , which is a VCLXAccessibleBox subclass,
and that is actually the VCLXAccessibleListItem's accessible
"grandparent" (parent's parent), as the direct parent is a
VCLXAccessibleList.
Therefore, the parent-relative position was incorrect.

Fix that by subtracting the parent's relative position
within the ListBox/ComboBox.

VCLXAccessibleListItem::getLocationOnScreen, which returns
the absolute screen position, was correct, because it
added the combobox/listbox position to the "parent-relative"
position (which was actually the position relative to the
combobox/listbox).
Now that VCLXAccessibleListItem::implGetBounds returns the
position relative to the actual parent, also adjust
VCLXAccessibleListItem::getLocationOnScreen to add the
actual parent's location on screen.

The issue about incorrect parent-relative positions
was noticed with an (upcoming) change that refactors
VCLXAccessibleListItem to subclass comphelper::OAccessibleComponentHelper
where the parent-relative position gets used.

Issue could also be reproduced like this:

1) apply only the VCLXAccessibleListItem::getLocationOnScreen
   part of this change

2) run Python script attachment 199141 from tdf#161087 on X11/XWayland
   which highlights the area of the currently focused object:

    GDK_BACKEND=x11 python3 ./highlight-focused-object.py

3) start LO with the qt6 VCL plugin on X11/XWayland:

    QT_QPA_PLATFORM=xcb SAL_USE_VCLPLUGIN=qt6 ./instdir/program/soffice.bin --writer

4) Navigate to the "Styles" combobox in the formatting toolbar

5) Press Alt+Down to expand, then arrow keys to navigate through
   the combobox entries

(Note: Running Orca in addition can help to trigger creation of the
the required a11y objects and events.)

Without this change in place, the wrong area was highlighted
when moving between entries. Now, the correct area of the currently
focused combobox entry is highlighted.

Change-Id: I53c686276537713fec604eb22d19be2331ce1f56
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181432
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>