list item a11y: Use comphelper::OAccessibleComponentHelper
Let VCLXAccessibleListItem subclass
comphelper::OAccessibleComponentHelper which already
implements the logic for event handling and most of the
location/bounds handling, instead of duplicating the
logic here.

The existing VCLXAccessibleListItem::implGetBounds
basically already implements the logic needed to override
the purely virtual
comphelper::OAccessibleComponentHelper::implGetBounds,
only the required conversion from tools::Rectangle to awt::Rectangle
is still added in this commit.

Tested as described in more detail in previous commit

    Change-Id: I53c686276537713fec604eb22d19be2331ce1f56
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 11 15:13:41 2025 +0100

        list item a11y: Fix parent-relative position

with no change in behavior noticed (now that the issues
detected previously have been addressed in previous
commits already in preparation for this commit).

Change-Id: I74e53e3297948977bed1b8608468dc8378859426
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181433
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>