ScriptForge (SF_Services) fix tdf#165147 do not require JRE
When a user script invokes CreateScriptService()
for a not yet used library, ScriptForge loads
the library and searches for the list of
services provided by that library.

The list can be found in a specific method
that the library must contain,
  RegisterScriptServices()
in whatever module.

The exploration was done so far through a
ScriptProvider UNO service. This service includes
all programming languages, including Java,
even without Java being used. This caused
an inconvenient error message when Java
is disabled in the advanced options.

This service is now replaced by a
  com.sun.star.comp.scripting.ScriptProviderForBasic
service. This gives a direct access to a lower
branch of the scripts tree and skips the Java issue.

All the changes are done in the
  ScriptForge.SF_Services._FindModuleFrom Method()
function.

No change required in the user documentation.

Change-Id: I4a6820e5d1a379bf5b61c6077db8a9a382c1757d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181445
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
(cherry picked from commit ff298dfebcd5a6772de53d075e960bddcd4aff46)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181482
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>