tdf#165207 tdf#164201 docx: always use errata uri in docProps/core.xml
This patch replaces my import-focused
25.8 commit d97085cc6cd2bdc3b6723d1960d0ec5fa0a48165 which said:
    This fixes a 7.6 regression
    from commit e66ddcd4b66923bc835bd7c5f5c784a809a420a2.

    At import, the base filter was treating too many documents
    as if they were limited to Word 2007 format,
    and thus reducing their compatibilityMode to 12 on export.

    This import case is matched in a LOT of unit tests.
    However, it doesn't manifest itself in
        saveAndReload(mpFilter) (or DECLARE_OOXMLEXPORT_TEST)
    because the mpFilter string set in the Test class with
        SwModelTestBase(..."ooxmlexport/data/", "Office Open XML Text"),
    forces saving in ISOIEC_29500_2008 mode
    and thus unit tests basically NEVER round-trip as "MS Word 2007 XML".

    However, the general user was almost always round-tripping
    these as MS Word 2007 XML / compat12 since LO 7.6.

The export portion of this patch
fixes a long-standing bug that became more prominent in 7.6
when mstahl renamed docx filters to promote Word 2010,
and then again in 25.2 when my bug 164201 patches
told command-line convert-to to export using Word 2010 filter,
and finally in 25.8 when interactively round-tripping existing files.

So what does this patch fix?
Basically, all of the meta-data was not being read by MS Word,
which primarily means no create/modify dates, subject, title etc.
Every time we saved to DOCX format (except DOCX 2007), we caused Word
to be incapable of understanding anything in docProps/core.xml.
Telltale sign: when MS Word round-trips with added docProps/core0.xml.

The spec says that DOCX 2010 should be using
    officedocuments/2006/relationships/metadata/core-properties
and thus our untouched 2010-era coding
specified all of that for import and export.

But reality seems to be much different than the spec,
(and some errata-sounding documentation agrees),
so throw the spec out the window and always write
    package/2006/relationships/metadata/core-properties
(but only for DOCX - Excel and Powerpoint follow the spec...)

In order to do that, I needed to re-write the import logic
as already suggested by mstahl in
https://gerrit.libreoffice.org/c/core/+/178048

Bug 165207 doesn't affect LO in any way.
If there was already a unit test writing out to _rels/.rels
I would have added a test for this, but since there isn't
I won't bother since I don't see any value in doing so.
Plus, we already have validity checks to cover this kind of stuff.

make CppunitTest_sw_ooxmlexport20 CPPUNIT_TEST_NAME=testTdf158855

A uiwriter4 test depends on the previous patch in this bug report.

officeotron sounds like they WANT to report invalidation errors
based only on spec, so no point in trying to "submit a patch".

Change-Id: Ib76803bf0c9f3791f9078846f00c118099ef67cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181463
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182037
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Justin Luth <jluth@mail.com>