tdf#165207 tdf#164201 docx: always use errata uri in docProps/core.xml
This patch effectively replaces (and thus contains a revert of)
25.8 commit d97085cc6cd2bdc3b6723d1960d0ec5fa0a48165.

This fixes a longstanding bug that became more prominent in 7.6
when mstahl renamed docx filters to promote Word 2010,
and then again in 25.2 when my bug 164201 patches
told command-line convert-to to export using Word 2010 filter,
and finally in 25.8 for interactive round-tripping of existing files.

So what does this patch fix?
Basically, all of the meta-data was not being read by MS Word,
which primarily means no create/modify dates, subject, title etc.
Every time we saved to DOCX format (except DOCX 2007), we caused Word
to be incapable of understanding anything in docProps/core.xml.
Telltale sign: when MS Word round-trips with added docProps/core0.xml.

The spec says that DOCX 2010 should be using
    officedocuments/2006/relationships/metadata/core-properties
and thus our untouched 2010-era coding
specified all of that for import and export.

But reality seems to be much different than the spec,
(and some errata-sounding documentation agrees),
so throw the spec out the window and always write
    package/2006/relationships/metadata/core-properties
(but only for DOCX - Excel and Powerpoint follow the spec...)

That meant that I also had to rewrite our import logic,
but that was already suggested by mstahl in
https://gerrit.libreoffice.org/c/core/+/178048

Bug 165207 doesn't affect LO in any way.
If there was already a unit test writing out to _rels/.rels
I would have added a test for this, but since there isn't
I won't bother since I don't see any value in doing so.
Plus, we already have validity checks to cover this kind of stuff.

The retained the unit test ensures import logic is still correct:
make CppunitTest_sw_ooxmlexport20 CPPUNIT_TEST_NAME=testTdf158855

A uiwriter4 test depends on the previous patch in this bug report.
(See earlier versions of my patchset for details)

officeotron sounds like they WANT to report invalidation errors
based only on spec, so no point in trying to "submit a patch".

Change-Id: Ib76803bf0c9f3791f9078846f00c118099ef67cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181463
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>