ruler a11y: Use comphelper::OAccessibleComponentHelper
Turn SvtRulerAccessible into a comphelper::OAccessibleComponentHelper
subclass.

comphelper::OAccessibleComponentHelper already implements
the logic for event handling and most of the location/bounds handling,
so avoid reimplmenting it in SvtRulerAccessible.

Turn SvtRulerAccessible::GetBoundingBox into the
`implGetBounds` override of the base class and let it return
an awt::Rectangle.

No change in behavior seen when testing with Accerciser.
(The class implements a11y e.g. for the ruler above the document
in Writer.)

Change-Id: I3e1f1dcd7533fb14317428e74b3be1a1e2e57277
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181477
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>