IASS: Use correct OutlinerParaObject in text decompositions
In SdrTextObj the methods impDecomposeAutoFitTextPrimitive and
impDecomposeChainedTextPrimitive used the wrong
OutlinerParaObject. Unfortunately there are two OPOs available
in the SdrTextPrimitive2D, one by using
    (A) getOutlinerParaObject()
and one using
    (B) getSdrText()->GetOutlinerParaObject()
whereby (A) is the one which contains possible changes from
an active text edit while (B) contains the original text at
the start of the possible text edit.
By using (B) those two decomposes were not able to show
the already changed text, thus the update in the SlideShow
seemed not to work.

Change-Id: Id2dcc3924fe8029688d0e76ca0f3f7300f461bd5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181490
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>