math a11y: Don't return self in SmGraphicAccessible::getAccessibleAtPoint
As the XAccessibleComponent::getAccessibleAtPoint doc says,
the purpose of this method is:

> Returns the Accessible child that is rendered under the given point.

Returning a reference to self is incorrect and bad.
It e.g. triggers an infinite loop when using Accerciser's
"Inspect object under mouse" feature when the mouse is
over the formula in Math, because it recursively queries
for the child at the position until null is returned.

Change-Id: I1aefd6fe69df32bb766351fe068d1c8b151156ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181496
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>