tdf#130857 qt weld: Support "Fontwork Gallery" dialog
Declare support for the "Fontwork Gallery" dialog  that can be
triggered like this:

1) start Writer
2) "Insert" -> "Fontwork"

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Currently, the icons in the dialog are very small.

Change-Id: Idbc6ba8c67becafa294e49b638e541add1b5bc21
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181509
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>