a11y: Introduce weld::IconView::set_item_accessible_name
Introduce a new weld::IconView::set_item_accessible_name
method that allows to set the accessible name for
an item in the IconView.

So far, IconView icons don't have any accessible
name if no text is set in addition to the icon.

There is some logic to use (or fall back to) the tooltip
text for the accessible *description* in both,
SalInstanceIconView (introduced in

    commit 2a28ebeef5ea3e2b01d836a7233d2316b765bf38
    Author: Mike Kaganski <mike.kaganski@collabora.com>
    Date:   Wed Jun 1 11:18:26 2022 +0300

        Accessibility for IconView

) and GtkInstanceIconView (introduced in

    commit ba0e36e607d1c380fd09b6725a4ebcb69ff399de
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Mar 9 11:38:36 2023 +0100

        tdf#153657 tdf#140659 gtk3 a11y: Use IconView item tooltip as a11y desc

), but having a way to directly set the a11y name
has the following advantages:

* Focusable objects should have an accessible name,
  not just an accessible description (that is meant
  to provide additional information only, and whether
  or not accessible descriptions are announced
  is usually configurable in screen readers).

* The tooltip text may or may not be appropriate as
  the accessible name.

* a11y name/description and tooltip are generally
  conceptually different things in UI toolkits, so
  reflect that in the weld API as well and set
  them explicitly rather than making assumptions
  in the implementations that need to implement the
  weld API using specific toolkits (VCL or third-party).

Therefore, introduce this method to allow explicitly
setting the accessible name and implement it for all
3 implementations (SalInstanceIconView, GtkInstanceIconView,
QtInstanceIconView).

It will be used to set a11y names in upcoming commits.
(The fallback to use the tooltip for the a11y description
can be dropped once an accessible name is set everywhere.)

For the VCL implementation (SalInstanceIconView), introduce
a new SvTreeListEntry::m_sAccesibleName member and prefer
that in AccessibleListBoxEntry::getAccessibleName if set.

For GtkInstanceIconView, the logic is similar to the one
in the existing
GtkInstanceIconView::set_item_accessible_description_from_tooltip
(which is gtk3-only, i.e. doesn't work for gtk4).

For QtInstanceIconView, set the Qt::AccessibleTextRole [1]
for the item.

[1] https://doc.qt.io/qt-6/qt.html#ItemDataRole-enum

Change-Id: I8fcbdddd74d75283469f046ff5b60f111051b021
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181616
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>