related tdf#164201 Word 2010 filter should be flagged as FILTER_DOCX
This backport includes dotx commit b4bc81e12533

And I have no idea what all the implications of this will be,
but a hint is to look for FILTER_DOCX
(even though the value OXML existed before FILTER_DOCX was defined)
or git grep -25 SfxFilter | grep -25 GetUserData

"OXML" was always in MS_WORD_2007-XML.xcu since it was created.
3.x commit 47c4912441c25438b4433069940a0dd4de385cca
Author: Oliver Bolte on Thu Jan 10 11:56:37 2008 +0000
    2007/02/26 11:10:48 os 1.1.8.1: file MS_Word_2007_XML.xcu

The one "GetUserData" implication I know
is that "Insert - Text from file"
uses a different code path and inserts the text with an additional CR.

For FILTER_DOCX,
likely extremely important is that it was not considered
IsInWriterfilterImport() before... which looks important for undo.
commit c4dab726caaa73be9f9c731312080143b0a0b89d
Author: Caolán McNamara on Thu Mar 12 12:43:10 2020 +0000
    ofz#21168 sw,writerfilter: limit writerfilter hack to writerfilter

FILTER_DOCX was introduced with this commit.
It seems to be irrelevant and harmless for "import function" of autotext
since DOCX files are already listed in the file picker...
5.4 commit 0b31912b1df719c5507b22a2f63086cb8921f597
Author: Szymon Kłos on Fri Apr 14 10:22:37 2017 +0200
    AutoText: .dotx files visible when import
[ACTUALLY AutoText: .docx/.docm/.dotx files visible when import]

Change-Id: I7c0b82bc2fb1730c8fff595b7171765339de0330
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181645
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181991
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Justin Luth <jluth@mail.com>