vcl a11y: No longer report IconView item tooltip as a11y desc
Now after

    Change-Id: I8fcbdddd74d75283469f046ff5b60f111051b021
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 13 13:42:05 2025 +0100

        a11y: Introduce weld::IconView::set_item_accessible_name

introduced API to explicitly set the a11y name for IconView
items and commits

    Change-Id: I59275d8020d0ebd46256bdc1531849cf5be5a353
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 13 13:48:35 2025 +0100

        svx a11y: Set accessible name for Fontwork Gallery items

    Change-Id: Ie54f41014822ec92576f30c72f569a8d186151c6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 13 15:00:43 2025 +0100

        sd a11y: Set a11y names for items in master page selector

    Change-Id: I74d1aa41896b70f6025fdc7e4784a2b6539ec6f6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 13 15:18:10 2025 +0100

        math a11y: Set a11y names for items in Elements sidebar deck

implemented setting a11y names for those items which previously
had no accessible name, but only an accessible description
based on the tooltip, drop the fallback of using the tooltip
for the accessible description of tooltip items that was originally
introduced in

        commit 2a28ebeef5ea3e2b01d836a7233d2316b765bf38
        Date:   Wed Jun 1 11:18:26 2022 +0300

            Accessibility for IconView

and which would now result in the same text being used for
the accessible name and accessible description for the
above cases when the VCL implementation (SalInstanceIconView) is used.
(Screenreaders usually still announce only one of them
if they are the same, but the accessible description is
no longer needed now.)

See commit message of the above-mentioned

    Change-Id: I8fcbdddd74d75283469f046ff5b60f111051b021
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 13 13:42:05 2025 +0100

        a11y: Introduce weld::IconView::set_item_accessible_name

for more details.

As far as I can see, the UI elements addressed by the
above commits are the only ones that relied on the
tooltip getting used, as other dialogs etc. using
IconView pass a string/text when calling
weld::IconView::insert etc., and the three ones above
are the only ones calling weld::IconView::connect_query_tooltip
to be able to provide a tooltip text this way.

In case I missed anything and this commit results in
IconView items no longer getting announced somewhere, an
accessible name should presumably be set for those as well,
using  weld::IconView::set_item_accessible_name (s. above
commits for examples).

Change-Id: I153e91292ec70bb3a6dc4e1d473dc0362af55c79
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181636
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins