tdf#130857 qt weld: Show IconView item tooltips
Register the QtInstanceIconView as an event filter
for the QListView that holds the items and show
the tooltip text for the currently hovered item,
if there is one.

This works fine for the "Insert" -> "Fontwork"
dialog in Writer with the qt6 VCL plugin and
SAL_VCL_QT_USE_WELDED_WIDGETS=1.

Change-Id: I9c6d0e3309d4303558d64e1ca83b636ac3de712e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181651
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>