configure: fix --disable-skia build of pdfium
A problem since commit 6007fbef5c2e4a0b80a4c0031ff36117aa41e63f (Update
pdfium to 7012, 2025-02-13), it's one thing that pdfium needs to be
built with clang-cl on Windows (similar to Skia), but our configure
didn't configure clang-cl when skia was off. So do it when either skia
or pdfium is enabled.

Reported at
<https://lists.freedesktop.org/archives/libreoffice/2025-February/093003.html>.

Change-Id: I261b98511b664182d46a4a11ecbb40ad91663c41
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181665
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins