cool#11064 vcl lok: expose info about the scheduler
A LOK client can interrupt the vcl main loop in its 'any input'
callback. When making that decision, it's useful in case it can know
what's the priority of the most urgent job, for example it may want core
to still finish high priority tasks, but not idle ones.

Add a LOK API to expose this info. Keep it minimal, so it's realistic to
call this frequently from a LOK client.

Change-Id: Id51668eb8156067e60d6fd0f33606c65858008a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181761
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins