tdf#130857 weld: Let SpinButton output hdl only format the string
Adjust the logic so that the weld::SpinButton::m_aOutputHdl
set via weld::SpinButton::connect_output is now only
responsible for returning a string format for a given spinbox
value, i.e. it now has a sal_Int64 input parameter and
a OUString return value.

This means that the implementations for those handlers
generally no longer need interact with the underlying
weld::SpinButton directly, i.e. no longer need to call
SpinButton::get_value and SpinButton::set_text to get the
value and manually set the formatted string as the new
text.

That logic is for now moved to
weld::SpinButton::signal_output which gets called by the
weld::SpinButton implementations and calls the handlers
(i.e. one level further up in the call stack) and other
places directly calling the handlers.

The main motivation is that in order to implement
setting the formatted text based on a value for QSpinBox/QDoubleSpinBox
can be done by overriding QDoubleSpinBox::textFromValue
that takes a value and returns a string. Changing
weld::SpinButton::m_aOutputHdl to a similar logic
is one step in preparation of implementing the logic
for QtInstanceSpinButton.

No change in behavior intended.

MetricSpinButton::spin_button_output was previously
explicitly checking whether the new and old text
were the same, and only calling weld::SpinButton::set_text
if they were not. If not doing that any more there should
result in any change in behavior, it might make sense
to reimplement that logic in the weld::SpinButton::set_value
implementations of the concrete weld::SpinButton subclasses
instead.

Change-Id: I53df480e5b3b2e4eeb9f9fb0c24e8e735fae063b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181684
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>