tdf#130857 weld: Let SpinButton::signal_output return text
If a custom output handler was set via
weld::SpinButton::connect_output, return the formatted string
it generates for the current value in SpinButton::signal_output
instead of directly setting it as the SpinButton's text using
SpinButton::set_text.

Let the subclasses which call the method take care of setting
the text instead.

If no handler is set, return an empty std::optional.

This approach is similar to how SpinButton::signal_input
returns the new value (via an out param).

But the main reason is that simply calling
weld::SpinButton::set_text to set the new text isn't sufficient
for the Qt implementation in QtInstanceSpinButton, but
QDoubleSpinBox::textFromValue  can be overriden for that
purpose - and this commit prepares for an implementation
that will do that in an upcoming commit.

Change-Id: I4962497122b79c320f3c60e5565cd64973af6df3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181685
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>