tdf#130857 weld: Let SpinButton input hdl only parse value from string
Similar to how

    commit bbc51c66a3b499c3f4ac71e4e6f7a0f2d9984dfc
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Feb 14 20:38:28 2025 +0100

        tdf#130857 weld: Let SpinButton output hdl only format the string

adjusted the logic to let weld::SpinButton::m_aOutputHdl only
format a value into a string, also adjust the logic for
weld::SpinButton::m_aInputHdl:

Let it take a string and convert that into a value/integer,
without having to retrieve the text from the SpinButton
itself.

As the text may not be parseable into an integer, use
std::optional<int> as a return value. (An empty std::optial means
that parsing failed, which was previously indicated by a return value
of `false`).

Besides simplifying the implementations (set via
SpinButton::connect_input), this also aligns the logic
more with what Qt's QDoubleSpinBox::valueFromText [1]
does and thus prepares for implementing the required
logic for QtInstanceSpinButton.

[1] https://doc.qt.io/qt-6/qdoublespinbox.html#valueFromText

Change-Id: Icb55f49cddeb1d039f4b4c8359058769cea77f14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181713
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins