sc: Don't modify string passed to NumToAlpha/AlphaToNum
The NumToAlpha and AlphaToNum helper functions
are only used by ScNavigatorDlg::ParseRowInputHdl
(and AlphaToNum also calls NumToAlpha in turn).

So far, both methods were taking an `OUString&`
param and modifying that string.

However, ScNavigatorDlg::ParseRowInputHdl doesn't
actually make use of the out param any more after
calling those methods, so switch the param to
`const OUString&` instead and stop modifying
the value in the helper functions.

In AlphaToNum, introduce a new local variable
`aUpperCaseStr` so that the modified string
can be used within that function without
modifying the string passed as a param.

In NumToAlpha, drop the call to ::ScColToAlpha,
whose only effect was to modify the out param,
and drop the now unused out param altogether.

Change-Id: Idcfd743603e12a06633b615aa8c912c244f0bd56
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181712
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins