tdf#130857 weld: Abstract from "input"/"output" SpinButton signals
So far, the weld::SpinButton logic to retrieve a value
from text in its entry and (the other way around) to format a value into
text to show in the entry was based on the concept of
GtkSpinButton and its "input" [1] and "output" [2] signals
for which handlers can be set that directly retrieve/set
the text and value in the SpinButton.

Continue and finish the work from previous commits

    Change-Id: I53df480e5b3b2e4eeb9f9fb0c24e8e735fae063b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Feb 14 20:38:28 2025 +0100

        tdf#130857 weld: Let SpinButton output hdl only format the string

and

    Change-Id: I4962497122b79c320f3c60e5565cd64973af6df3
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Feb 15 01:19:56 2025 +0100

        tdf#130857 weld: Let SpinButton::signal_output return text

and

    Change-Id: Icb55f49cddeb1d039f4b4c8359058769cea77f14
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Feb 15 11:22:54 2025 +0100

        tdf#130857 weld: Let SpinButton input hdl only parse value from string

to abstract away from those signals.

Consider this as an implementation detail that dialog
implementations don't have to deal with.
It also doesn't match the approach that Qt's QSpinBox/QDoubleSpinBox
uses for the value <-> text conversion.

Instead, use the approach of having methods that
simply convert a given value to a formatted string
and vice versa, and call those from the "output" callback/handler
in the GTK implementation of weld::SpinButton, GtkInstanceSpinButton,
and call the method to set text/value with the returned value there.

Dialogs (users of weld::SpinButton) can set a "value formatter"
(a method that returns a formatted string for a given value) and a
"text parser" (that returns an integer value for a given
text).

Part of the logic was already introduced in the above-mentioned
commits. This commit adjusts the former weld::SpinButton::signal_output
to take the value to format as an input parameter instead
of calling weld::SpinButton::get_value itself to get the
current value.
Similarly, the former weld::SpinButton::signal_input
now gets the text to parse as an input paramter instead
of calling weld::SpinButton::get_text itself.

Rename the methods to reflect the new semantics:

* weld::SpinButton::signal_output -> weld::SpinButton::format_value
* weld::SpinButton::signal_input -> weld::SpinButton::parse_text
* weld::SpinButton::connect_output -> weld::SpinButton::set_value_formatter
* weld::SpinButton::connect_input -> weld::SpinButton::set_text_parser

GtkInstanceSpinButton and SalInstanceSpinButton are adjusted
to pass the values accordingly, so no change in behavior
is to be expected.

[1] https://docs.gtk.org/gtk3/signal.SpinButton.input.html
[2] https://docs.gtk.org/gtk3/signal.SpinButton.output.html

Change-Id: If4cd12ef947233331a5deccba5c14a15e30a30c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181714
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>