svdraw: append SdrCropViewHdl instead of prepending in crop mode
Previously, enabling crop mode caused the crop handle (SdrCropViewHdl) to be added
at the start of the handlers list. This change in order led to a shift in handler IDs,
which Collabora Online relies on for identifying and managing handlers.

This patch fixes the bug by appending the SdrCropViewHdl to the end of the list,
ensuring that the IDs for the main crop handles remain consistent.

Change-Id: If082d292e32e242648cf5d04e278880bd457efb5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181705
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Gökay ŞATIR <gokaysatir@collabora.com>
(cherry picked from commit 4daf1b96c52e8b399908bfeb6871482a6699422c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181775
Tested-by: Jenkins