tdf#130857 qt weld: Implement SpinButton text <-> value conversion
In order to implement custom conversion between the current
value of a spin box and the text to display as well as from
the user-entered text to the value to use, methods
QDoubleSpinBox::textFromValue [1] and
QDoubleSpinBox::valueFromText [2] can be overriden.

Implement these methods and use the existing methods
in QtInstanceSpinButton's weld::SpinButton base class
to implement the logic.

In order to be able to use those in QtDoubleSpinBox,
add two new members and setters for those functions,
and call them in the QtInstanceSpinButton ctor.

See also previous commit

    Change-Id: If4cd12ef947233331a5deccba5c14a15e30a30c8
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Feb 15 13:44:50 2025 +0100

        tdf#130857 weld: Abstract from "input"/"output" SpinButton signals

and the commits mentioned therein for more background,
as those commits were in preparation for this one.

In a WIP branch adding support for the "Format" -> "Font Size"
dialog in Math, this commit makes the text in the
spin boxes show the units as expected, e.g. "12 pt", "100 %"
when using SAL_VCL_QT_USE_WELDED_WIDGETS=1.

(Support for that dialog will be declared in an upcoming
commit.)

[1] https://doc.qt.io/qt-6/qdoublespinbox.html#textFromValue
[2] https://doc.qt.io/qt-6/qdoublespinbox.html#valueFromText

Change-Id: Ifb24da2bcf37c2a769d22d5aa33b69929ca8cb11
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181715
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins