tdf#130857 qt weld: Demote assert in QtInstanceEntry::set_width_chars
Switch the assert that the method isn't implemented to a SAL_WARN
for now, to not block other Qt welding work at the moment
(like adding support for the Math "Format" -> "Font Size" dialog).

This method is for adjusting the size of the widget, so not having
it implemented should have "only" a visual effect, while at this
stage, my primary focus is on implementing missing logic and
dialogs are added one at a time, so any major issues will be
noted while working on adding support for them.

Come back to this later.

Change-Id: I65bed22716db92a04da2a41dc1966ab5635b0482
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181716
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>