tdf#130857 qt weld: Support Math "Font Size" dialog
Declare support for the "Format" -> "Font Size" dialog in
Math.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This dialog uses weld::MetricSpinButton, and is therefore
the first supported dialog making use of the logic implemented in

    Change-Id: Ifb24da2bcf37c2a769d22d5aa33b69929ca8cb11
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Feb 15 17:47:01 2025 +0100

        tdf#130857 qt weld: Implement SpinButton text <-> value conversion

Change-Id: Idaaa07ae3da884436af4b97fc06886098db45936
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181717
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins