tdf#130857 qt weld: Split QtBuilder::setProperties
Split QtBuilder::setProperties into 4 more specific
methods that handle particular widgets and call the
methods right after the widgets have been created
and no longer from QtBuilder::insertObject.

This aligns this with how its already done for
other widget types and removes the need to
use qobject_cast, as the class is known at the
point in time when the specific widgets are created.

Change-Id: I230c07b97ea15cdc83ba0965f36496b930323860
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181723
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins